requirements.txt
absl-py==2.1.0
altair==5.2.0
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
awscli==1.22.34
Babel==2.8.0
bcrypt==3.2.0
blinker==1.8.2
botocore==1.23.34
branca==0.7.1
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
# Removed cloud-init==23.3.3 due to dependency issue
colorama==0.4.3
# Removed command-not-found==0.3 due to dependency issue
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
# Removed distro-info==1.1+ubuntu0.1 due to dependency issue
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
# Removed ec2-hibinit-agent==1.0.0 due to dependency issue
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
folium==0.15.1
fonttools==4.48.1
fsspec==2024.5.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
h5py==3.11.0
# Removed hibagent==1.0.1 due to dependency issue
httplib2==0.20.2
huggingface-hub==0.23.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
# Removed keras==3.3.3 due to dependency issue
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
# Removed matplotlib==3.8.2 due to dependency issue
mdurl==0.1.2
# Removed ml-dtypes==0.3.2 due to dependency issue
more-itertools==8.10.0
namex==0.0.8
netifaces==0.11.0
nltk==3.8.1
# Removed numpy==1.26.4 due to dependency issue
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
# Removed pandas==2.2.0 due to dependency issue
pexpect==4.8.0
pillow==10.2.0
playwright==1.41.1
protobuf==4.25.2
ptyprocess==0.7.0
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydeck==0.8.1b0
pyee==11.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
# Removed python-apt==2.4.0+ubuntu2 due to dependency issue
# Removed python-debian==0.1.43+ubuntu1.1 due to dependency issue
python-docx==1.1.2
python-dotenv==0.19.2
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
regex==2024.5.15
requests==2.31.0
rich==13.7.0
roman==3.3
rsa==4.7.2
s3transfer==0.5.0
safetensors==0.4.3
# Removed scikit-learn==1.4.0 due to dependency issue
# Removed scipy==1.12.0 due to dependency issue
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
# Removed sos==4.5.6 due to dependency issue
ssh-import-id==5.11
streamlit==1.31.0
# Removed systemd-python==234 due to dependency issue
tenacity==8.2.3
# Removed tensorboard==2.16.2 due to dependency issue
tensorboard-data-server==0.7.2
# Removed tensorflow==2.16.1 due to dependency issue
# Removed tensorflow-io-gcs-filesystem==0.37.0 due to dependency issue
termcolor==2.4.0
texttable==1.6.4
# Removed tf_keras==2.16.0 due to dependency issue
threadpoolctl==3.2.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
transformers==4.41.0
Twisted==22.1.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
# Removed ubuntu-advantage-tools==8001 due to dependency issue
# Removed ufw==0.36.1 due to dependency issue
urllib3==1.26.5
validators==0.22.0
wadllib==1.3.6
watchdog==4.0.0
websocket-client==0.59.0
Werkzeug==3.0.3
wrapt==1.16.0
xyzservices==2023.10.1
zipp==1.0.0
zope.interface==5.4.0
openai==0.27.0
flask-cors==3.0.10
